Changed mabs timer to treat MIN_POLL_TIME_US as microseconds instead of 10's of nanoseconds
The mabs button and LED server was updating the LEDs every MIN_POLL_TIME_US reference clock ticks. But, the macro name implies that it should be updating the LEDs every MIN_POLL_TIME_US microseconds, rather than reference clock ticks. So, this fixes that. This also fixes an issue with the BGA167 boards where the MABS server is broken when the core clock is 500 MHz, because the timer is firing so rapidly it's never able to service any other events.